divide SelectedFilters into parent and child component
